Update events to only include relevant events
Events can be limited to specific events. To record the `paid` status we
only need to listen to the invoice paid event. The invoice of the event
type `invoice.paid` is always paid.

For subscriptions we need to record the `customer.subscription.created` event
to create the initial record. The only relevant status it `canceled`. The
`customer.subscription.deleted` return subscription records with status
`canceled`.